Create helpers.py and func to check if pdf is encrypted
